requirements.txt
requests
bs4
Django==3.1.4
djangorestframework==3.12.2